Changed a casting
git-svn-id: https://code.nr.no/scm/svn/crava/trunk@206 176de4d3-4747-0410-aa96-47f6d6634214

+++ src/model.cpp
@@ -612,7 +612,7 @@ Model::makeTimeSimboxes(Simbox        *& timeSimbox,
   failed = true;
   }
   // Rotate variograms relative to simbox
-  modelSettings_->rotateVariograms(timeSimbox_->getAngle());
+  modelSettings_->rotateVariograms(static_cast<float> (timeSimbox_->getAngle()));
 
   //
   // Set SURFACES
